Combine FoldersPicker and Modal
